backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
requests==2.32.3
beautifulsoup4==4.12.3
pdfplumber==0.11.4
openai>=2.0.0
python-dotenv==1.0.1
tenacity==9.0.0
